+++ CHANGELOG.md
@@ -1,13 +1,59 @@
-## [1.0.1] - before null-safe version 
-## [1.0.2] - fix widget doesn't transform
-## [1.0.3] - remove unnecessary task
+## 3.0.1 
 
-## [2.0.0] - adapter null-safe version 
-## [2.0.1] - modify notification usage
-## [2.0.2] - fix widget doesn't transform
-## [2.0.3] - remove unnecessary task
-## [2.0.4] - improve code style
-## [2.0.5] - add export file
-## [2.0.6] - format code
+* add the `builder` to create a widget when it's need and the `child` is deprecated.
 
-## [3.0.0] - adapted 3.0
+## 3.0.0
+
+* adapted 3.0
+
+## 2.0.6 
+
+* format code
+
+## 2.0.5
+
+* add export file
+
+## 2.0.4
+
+* improve code style
+
+## 2.0.3 
+
+* remove unnecessary task
+
+## 2.0.2 
+
+* fix widget doesn't transform
+
+## 2.0.1 
+
+* modify notification usage
+
+## 2.0.0 
+
+* adapter null-safe version
+
+## 1.0.3 
+
+* remove unnecessary task
+
+## 1.0.2 
+
+* fix widget doesn't transform
+
+## 1.0.1
+
+* before null-safe version 
+
+
+
+ 
+
+
+
+
+
+
+
+